+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Braintree iOS SDK Release Notes
 
+## unreleased
+* American Express
+  * Remove unused `BTAmericanExpressErrorDomain`
+  * Remove unused `BTAmericanExpressErrorType`
+
 ## 5.5.0 (2021-11-01)
 * Add `displayName` to `BTLocalPaymentRequest`
 * Add `riskCorrelationId` to `BTPayPalRequest`